feat: 드롭 존을 inputbox에서 chat 전체 범위로 이동, chat 메서드 provider를 컨테이터에서 감싸도록 변경
