+++ next/src/components/video-chat/ChatSection/Container/index.tsx
@@ -1,8 +1,9 @@
 import React, { ReactNode, useEffect, useState } from 'react';
 import { ChatEvent } from '@lib/enum';
 import useChatClient from '@hooks/chat/useChatClient';
-import type { ChatSectionProps } from '..';
 
+import useDropzone from './useDropzone';
+import type { ChatSectionProps } from '..';
 import * as S from './style';
 
 export interface ContainerProps extends ChatSectionProps {
@@ -12,6 +13,7 @@ export interface ContainerProps extends ChatSectionProps {
 const Container = ({ children, isShowChat, setIsShowChat }: ContainerProps) => {
   const [memberCount, setMemberCount] = useState<number>(0);
   const socketClient = useChatClient();
+  const { getRootProps, getInputProps } = useDropzone();
 
   const handleChatCloseButtonClick = () => {
     setIsShowChat(false);
@@ -28,14 +30,17 @@ const Container = ({ children, isShowChat, setIsShowChat }: ContainerProps) => {
   }, [socketClient]);
 
   return (
-    <S.Container isShow={isShowChat}>
-      <S.Header>
-        <S.ChatCloseButton onClick={handleChatCloseButtonClick} />
-        <S.ChatTitle>채팅</S.ChatTitle>
-        <S.ChatMemberCount>{memberCount}</S.ChatMemberCount>
-      </S.Header>
-      {children}
-    </S.Container>
+    <div {...getRootProps()}>
+      <input {...getInputProps()} />
+      <S.Container isShow={isShowChat}>
+        <S.Header>
+          <S.ChatCloseButton onClick={handleChatCloseButtonClick} />
+          <S.ChatTitle>채팅</S.ChatTitle>
+          <S.ChatMemberCount>{memberCount}</S.ChatMemberCount>
+        </S.Header>
+        {children}
+      </S.Container>
+    </div>
   );
 };
