+++ next/src/pages/video-chat/[id].tsx
@@ -9,7 +9,6 @@ import CustomHead from '@components/common/CustomHead';
 import Container from '@components/video-chat/Container';
 import CamSection from '@components/video-chat/CamSection';
 import ChatSection from '@components/video-chat/ChatSection';
-import { ChatProvider } from '@components/video-chat/ChatSection/InputBox/ChatContext';
 import Sidebar from '@components/video-chat/Sidebar';
 import apiClient, { logAxiosError } from '@lib/apiClient';
 import devModeLog from '@lib/devModeLog';
@@ -63,12 +62,7 @@ const ChatRoom = () => {
       <SocketContext.Provider value={socketClient}>
         <Container>
           <Sidebar setIsShowChat={setIsShowChat} />
-          <ChatProvider>
-            <ChatSection
-              isShowChat={isShowChat}
-              setIsShowChat={setIsShowChat}
-            />
-          </ChatProvider>
+          <ChatSection isShowChat={isShowChat} setIsShowChat={setIsShowChat} />
           <CamSection />
         </Container>
       </SocketContext.Provider>
